Move IsNil back into utils
