+++ utils/misc.go
@@ -1,0 +1,22 @@
+package utils
+
+import (
+	"reflect"
+)
+
+// IsNil returns whether the given object is nil or an interface to a nil
+func IsNil(v interface{}) bool {
+	// if v doesn't have a type or value then v == nil
+	if v == nil {
+		return true
+	}
+
+	val := reflect.ValueOf(v)
+
+	// if v is a typed nil pointer then v != nil but the value is nil
+	if val.Kind() == reflect.Ptr {
+		return val.IsNil()
+	}
+
+	return false
+}
